+++ docs/markdown/progress-notes/changelog.md
@@ -201,3 +201,5 @@
 - Treating a float as a 64-bit integer now sign extends, like on a real PS2
 - It is now an error to have two arguments with the same name.
 - It is now a warning to redefine a constant.
+- Fix a bug where the size of static boxed arrays was only `length` and not `allocated-length`
+- It is now possible to call a method on a forward declared type. The forward declared type must be a basic.